fix problem with travis cache If sudo changed then cache has different signature

+++ travis/gitupdf.sh
@@ -7,45 +7,42 @@
 #
 
 if [ "$TRAVIS_BRANCH" = "master" ]; then
-    case "$OWTRAVISJOB" in
-        BOOTSTRAP|BUILD|DOCPDF)
-            if [ "$TRAVIS_EVENT_TYPE" = "push" ]; then
-                cd $OWRELROOT
-                #
-                # remove all local changes to upload only error logs
-                #
-                git reset --hard
-                git clean -fd
-                #
-                # copy build log files to git repository tree
-                #
-                if [ "$TRAVIS_OS_NAME" = "osx" ]; then
-                    if [ ! -d bld/osx ]; then mkdir -p logs/osx; fi
-                    cp $TRAVIS_BUILD_DIR/bld/*.log logs/osx/
-                else
-                    if [ ! -d bld/linux ]; then mkdir -p logs/linux; fi
-                    cp $TRAVIS_BUILD_DIR/bld/*.log logs/linux/
-                fi
-                #
-                # commit new log files to GitHub repository
-                #
-                git add -f .
-                if [ "$TRAVIS_OS_NAME" = "osx" ]; then
-                    git commit --quiet -m "Travis CI build $TRAVIS_JOB_NUMBER (failure) - log files (OSX)"
-                else
-                    git commit --quiet -m "Travis CI build $TRAVIS_JOB_NUMBER (failure) - log files (Linux)"
-                fi
-                git push --quiet -f origin
-                cd $TRAVIS_BUILD_DIR
-                echo "gitupdf.sh - done"
+    if [ "$OWTRAVISJOB" = "BOOTSTRAP" ] || [ "$OWTRAVISJOB" = "BUILD" ] || [ "$OWTRAVISJOB" = "DOCPDF" ]; then
+        if [ "$TRAVIS_EVENT_TYPE" = "push" ]; then
+            cd $OWRELROOT
+            #
+            # remove all local changes to upload only error logs
+            #
+            git reset --hard
+            git clean -fd
+            #
+            # copy build log files to git repository tree
+            #
+            if [ "$TRAVIS_OS_NAME" = "osx" ]; then
+                if [ ! -d bld/osx ]; then mkdir -p logs/osx; fi
+                cp $TRAVIS_BUILD_DIR/bld/*.log logs/osx/
             else
-                echo "gitupdf.sh - skipped"
+                if [ ! -d bld/linux ]; then mkdir -p logs/linux; fi
+                cp $TRAVIS_BUILD_DIR/bld/*.log logs/linux/
             fi
-            ;;
-        *)
+            #
+            # commit new log files to GitHub repository
+            #
+            git add -f .
+            if [ "$TRAVIS_OS_NAME" = "osx" ]; then
+                git commit --quiet -m "Travis CI build $TRAVIS_JOB_NUMBER (failure) - log files (OSX)"
+            else
+                git commit --quiet -m "Travis CI build $TRAVIS_JOB_NUMBER (failure) - log files (Linux)"
+            fi
+            git push --quiet -f origin
+            cd $TRAVIS_BUILD_DIR
+            echo "gitupdf.sh - done"
+        else
             echo "gitupdf.sh - skipped"
-            ;;
-    esac
+        fi
+    else
+        echo "gitupdf.sh - skipped"
+    fi
 elif [ "$COVERITY_SCAN_BRANCH" = 1 ]; then
     echo "gitupdf.sh - skipped"
 else
